added admm test and refactorings
